Major performance increase: hash maps
More or less the same data structure, as far as memory
usage is concerned: a single array containing all the
countries, once each.

But instead of binary search, we generate a hash table
that indexes ALL available property values and points
them at the object.

There is no change in the public interface.